app/requirements.txt
flask
flask_wtf
mysql-connector-python
pycryptodome
passlib
waitress
argon2-cffi